Class MyMath renamed
